requirements.txt
discord.py
gspread
google-api-python-client
pytz
pillow
python-dateutil
pynacl